Use pathname as schema key name FOLIO-573 RMB-30
When any schema file refers to an additional schema file,
then use the pathname of that referenced second schema
as the "key" name in the RAML "schemas" section, and
wherever that schema is utilised in RAML and schema files.
Also ensure that all such referenced files are below the parent file.