remove multipart form data for now
seems like this is generated into a pojo with references to java.io.File objects. the actual content of the files is not part of the generated pojo so this seems useless. passing Files from client to server where the actual content is not part of the payload is useless