Allow recusion in JSON schemas references (loops) RMB-265
Instead of statically including JSON content for $ref, we just make
a pointer that raml-to-jaxrs can work with (that is all JSON schemas
in target/classes has that file:path instead of a relative refernence
that raml-cop is happy with. Note that raml-cop chokes on loops, which
is kind of weird since loops should be quite common in the description
of data structures.