RMB-468: Foreign key field index
Always create an index for the foreign key table field that has been extracted from
the jsonb. This UUID index is smaller and faster than an index that extracts the UUID
string from the josnb object.

Change all queries to use that field and index, for example use users.groupId in favour
of users.jsonb->'groupId'.