RMB-462: Remove the schema.json fullText / defaultDictionary option, always use 'simple'.
Always use 'simple' as the full text dictionary.

Remove the fullText defaultDictionary option from schema.json that is not documented,
used by mod-inventory-storage only, and does not work because cql2pg doesn't use it.

Also no longer use the x-okapi-language header in the postTenant method because this value
is not stored and is not used by cql2pg.