RMB-662: Remove parseDatabaseSchemaString and org.json:json
CQL2PGCLIMain.parseDatabaseSchemaString is an internal
method (package-private) that is not used.

The package dependency org.json:json violates Apache 2.0
license because it has this incompatible license requirement:
"The Software shall be used for Good, not Evil."
https://github.com/stleary/JSON-java/blob/20200518/LICENSE

If anyone needs the functionality of the method
parseDatabaseSchemaString that is only 25 lines long it
should be re-implemented using Vert.x JsonObject or
jackson-databind.

Tasks:
* Remove method CQL2PGCLIMain.parseDatabaseSchemaString
* Remove package org.json:json from dependencies.