RMB-671: Reinstall PostgreSQL extensions on upgrade
As an SysOp I want to run the upgrade on a system that is based on a freshly
installed backup. The backup may miss the PostgreSQL extensions that
reside in public schema.
Example: https://issues.folio.org/browse/MODORDSTOR-161

To make the life of SysOps easier be forgiving and reinstall all PostgreSQL
extensions on upgrade.