RMB-687: ResponseException for TenantAPI
Use a ResponseException that contains a javax.ws.rs.core.Response
when a TenantAPI call fails.

The TenantAPI methods use a Handler<AsyncResult<Response>> handler
to return success and failure. It is returned as a succeeded
AsyncResult where the value is the Response:
https://github.com/folio-org/raml-module-builder/blob/v30.2.4/domain-models-runtime/src/main/java/org/folio/rest/impl/TenantAPI.java

On success the Response has a 200 or 201 HTTP status code.
If installation or upgrade fails the Response has a 400 or 500
HTTP status code and a detailed error message in the body.
Note that AsyncResult is a succeeded result in all cases.

The documentation was wrong in that it only checked the
AsyncResult (succeeded, failed) and not the Response status
code.

This is how we fix this:
A success is returned using a succeeded AsyncResult as before.
A failure is returned using a failed AsyncResult. A failed AsyncResult
doesn't have a value where we can pass the Response. But it has an
Exception and we can create our custom ResponseException that contains
the Response.

RestVerticle extracts the Response from the ResponseException.