RMB-711: Remove Java 8 and Maven 3.3 check
We no longer need to check that the Java version is >= 1.8.0-101 because
we've already switched to Java 11 and both Maven and the Java runtime fail
with a good error message when running with Java 8.

We no longer need to detect a maven version older than 3.3 because such
versions are from 2014-12-20 or before and are unlikely to be still in use.