Fix NPE in HTTPJsonResponseHandle RMB-780
The close of client which was previously handled in exceptionHandler
(in RMB 31 series).. Do not close it in all cases. Do not check for
"Accept" header. It's useless, because there's no Accept header
in response headers!

Add a test for the case where response can not be encoded as JSON.