Relax on Accept and Content-Type RMB-519
Accept default is "*/*". Content-Type default is "application/json".
The existing implementation hard-coded content-types for which
RMB would buffer the incoming request. This is now simplified and
no longer hard-coded (and default operation is to buffer the
incoming request). So this fixes the problem for RMB support all
content types for regulare operations.