RMB-483 testContainers for testing with Postgres (#846)
This is an extension to RMB that allows Postgres testing with TestContainers - https://issues.folio.org/browse/RMB-483.

The existing postgres embedder is still enabled by default if "host" config is absent or if embed_postgres=true on the command-line. This PR makes RMB create a new artifact postgres-testing that should be included for modules in scope test. A module enables container based testing with:

  PostgresClient.setPostgresTester(new PostgresTesterContainer());

Next major RMB could go one step further and not the include old tester. This would be a trivial, but will require modules to include postgres-testing artifact and enable postgres containers with the above construct.

Co-authored-by: julianladisch <julianladisch@users.noreply.github.com>