RMB-849: Remove duplicate ModuleName class from domain-models-api-interfaces
domain-models-api-interfaces (and domain-models-runtime
that depends on it) ships with a ModuleName class with
getModuleName() returning "raml_module_builder".

An RMB based module creates a correct ModuleName class
when running domain-models-maven-plugin.

If we are lucky the latter is included in the final jar.
For mod-feesfines I sometimes have integration tests failure
because of wrong order resulting in the
wrong "raml_module_builder" where it should be "mod_feesfines".

Removing ModuleName class from domain-models-api-interfaces jar
resolves this issue; it requires reflection to load
ModuleName class, though.