RMB-886: TenantLoading withPostIgnore(), no overwrite on upgrade
When upgrading a module using TenantLoading it might be desired not
to overwrite existing data. When installing a module for a tenant loading
reference and sample data should be idempotent so that it can be run
twice, for example after a failure.

Solution: Add withPostIgnore() to TenantLoading that ignores
any 4xx HTTP status codes.