Perform module check at compile time RMB-903
The resulting schema in RMB has a limit of 63 characters
as that is the default identifier max length for Postgres.
However, if the tenant is know to be at most 31 bytes it's
safe to asseume the module may be up to 31 bytes. This PR
checks the module name during build time. An invalid name would
result in an errors such as:
[INFO] Reading from pom.xml
[ERROR] Invalid module name "007" does not match ^[a-z][-_a-z0-9]*$
java.lang.IllegalArgumentException: Invalid module name "007" does not match ^[a-z][-_a-z0-9]*$
    at org.folio.rest.tools.PomReader.readIt (PomReader.java:91)
    at org.folio.rest.tools.PomReader.init (PomReader.java:54)