Works on mac and linux now, but with caveats
The network host is now x platform. Needed container port was all. When pg_basebackup creates
the files in the temp dir on the host it sets very restrictive perms on the files. This means on
linux that we can't use the cleaner withCopyFileToContainer method when creating the final
standby. Instead wtihFileSystemBind works (no perms issue with it). And on linux there appears
to be no delay between writing and reading with this setup. But there is on mac.