Makes streaming synchronous
This fixes the problem with the delay when using withFileSystemBind on the mac -- a sign that
async replication (the default) would likely make even the simplest tests very flaky since
any verification of writes could fail in unexpected ways due to the slight delay.

With this change the sync_state column of pg_stat_replication now reports sync and
the immediate read on the standby returns a row.