RMB-973: If CQL is invalid PgUtil.getStream should log and report the CQL query
If the CQL query passed into PgUtil.getStream cannot be parsed the response shows the message and the class name of the exception (and the stack trace):

```
10:06:02 [922744/item-storage] [diku] [ce15840c-de2d-4fa3-a5bd-9d7e35b0e8b2] [mod_inventory_storage] ERROR PgUtil org.folio.cql2pgjson.exception.QueryValidationException: org.z3950.zing.cql.CQLParseException: expected index or term, got ')'
```

However, the CQL query is not shown. This makes it difficult to debug the issue.

Solution: Append the CQL query to the error message.