New caching approach
Cache is no longer fixed size and is synchronized with the underlying pool. Sessions for tenant can be reused
from the cache until they cycle out. A tenant can have more than one session available to it from the
cache.

Cached connections are only given out when they are available.

When the underlying pool is exhausted we remove connections from the cache when they are available.