RMB-360: db init without superuser/create rule/create schema privileges
https://folio-org.atlassian.net/browse/RMB-360

Some sysops want to give the module's database user no superuser, no
"CREATE ROLE" and no "CREATE SCHEMA" privileges and create the role
and the schema beforehand, before the module gets enabled for the
tenant.

Change RMB to gracefully handle this case.

This code has be successfully tested with mod-permissions.