RMB-1023: Remove/Reduce Ineffective Log Lines For CQL2PgJSON
https://folio-org.atlassian.net/browse/RMB-1023

These log lines could be up to 90% of logs generated by a module within a day.
This makes FOLIO modules logs more costly to store and adds to the amount of logs that need to be filtered out when troubleshooting.
The log lines are not as useful in production scenarios unlike development.
Reduce the log level of the following lines to DEBUG.

(cherry picked from commit de72e831188a8d315524b1857e249f7978b09b23)